Refactor: Update calls to moved OutputLogger
OutputLogger was moved to the common lib directory, update all
references.
