+++ server/src/ContextHandler.ts
@@ -24,7 +24,7 @@ import type {
   SymbolScanner
 } from './SymbolScanner'
 
-import logger from 'winston'
+import { logger } from './lib/src/utils/OutputLogger'
 import bitBakeProjectScanner from './BitBakeProjectScanner'
 
 /**
