+++ server/src/OutputParser.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import logger from 'winston'
+import { logger } from './lib/src/utils/OutputLogger'
 
 import {
   DiagnosticSeverity,
