+++ server/src/connectionHandlers/onCompletion.ts
@@ -8,7 +8,7 @@
  * Reference: https://github.com/bash-lsp/bash-language-server/blob/8c42218c77a9451b308839f9a754abde901323d5/server/src/server.ts#L408
  */
 
-import logger from 'winston'
+import { logger } from '../lib/src/utils/OutputLogger'
 import { type TextDocumentPositionParams, type CompletionItem, type SymbolInformation, CompletionItemKind } from 'vscode-languageserver/node'
 import { symbolKindToCompletionKind } from '../utils/lsp'
 import { BITBAKE_VARIABLES } from '../completions/bitbake-variables'
