+++ server/src/connectionHandlers/onDefinition.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import logger from 'winston'
+import { logger } from '../lib/src/utils/OutputLogger'
 import { type TextDocumentPositionParams, type Definition } from 'vscode-languageserver/node'
 import { analyzer } from '../tree-sitter/analyzer'
 import contextHandler from '../ContextHandler'
